fcm working and change of package name.

+++ app/src/main/java/com/teamx/hatlyDriver/data/dataclasses/totalEarning/SpentTime.kt
@@ -1,6 +1,6 @@
 package com.teamx.hatlyDriver.data.dataclasses.totalEarning
 
 data class SpentTime(
-    val hours: Int,
-    val minutes: Int
+    val hours: Double,
+    val minutes: Double
 )
