+++ app/src/main/java/com/teamx/hatlyDriver/data/dataclasses/totalEarning/TotalEarningData.kt
@@ -3,7 +3,7 @@ package com.teamx.hatlyDriver.data.dataclasses.totalEarning
 data class TotalEarningData(
     val riderDetail: RiderDetail,
     val spentTime: SpentTime,
-    val totalEarning: Int,
-    val totalOrder: Int,
-    val totalParcel: Int
+    val totalEarning: Double,
+    val totalOrder: Double,
+    val totalParcel: Double
 )
